+++ packages/frameworks/vue/src/presence/presence-props-context.ts
@@ -2,7 +2,7 @@ import type { ComputedRef } from 'vue'
 import { createContext } from '../context'
 import type { UsePresenceProps } from './use-presence'
 
-export type PresencePropsContext = ComputedRef<UsePresenceProps>
+export interface PresencePropsContext extends ComputedRef<UsePresenceProps> {}
 
 export const [PresencePropsProvider, usePresencePropsContext] =
   createContext<PresencePropsContext>('PresencePropsContext')