+++ packages/frameworks/vue/CHANGELOG.md
@@ -17,11 +17,14 @@ description: All notable changes to this project will be documented in this file
 
 - Revised `Dialog` component stories and tests
 - Revised `HoverCard` component stories and tests
+- Revised `Menu` component
 - Revised `Popover` component stories and tests
+- Revised `Select` component stories and types
 - Revised `Splitter` component
 - Revised `Tooltip` component
 - Revised `Tabs` component types
 - Revised `TagsInput` component
+- Revised `Toast` component
 - Revised `ToggleGroup` component
 
 ### Fixed
@@ -29,7 +32,11 @@ description: All notable changes to this project will be documented in this file
 - Added a missing slot with machine api in `Splitter`
 - Added a missing slot with machine api in `Tabs`
 - Added a missing slot with machine api in `Tooltip`
+- Added missing emits in `Dialog`
+- Added missing emits in `Presence`
+- Added missing emits in `Slider`
 - Fixed `Combobox.ItemText` type from `button` to `span`
+- Fixed `Toast` render bugs
 - Resolved an issue where `Tooltip` component would not handle open control state
 
 ## [0.9.0] - 2023-12-14